fix: improve usage of Ignore Pain
Suggest Ignore Pain as a high priority if Seeing Red is at 7 stacks or
if we should dump Rage to avoid capping. This removes the need to use
Revenge or Execute to dump Rage as they are on the GCD.

This changes from before where Ignore Pain was only suggested if
actively taking damage, but that's not necessarily needed due to Anger
Management.