fix: update to use additional settings from Hekili
Use new settings added to Hekili for protection warriors by noting them
as *tunables* in the priority.

Add an opener priority to try to capture the initial Rage and threat
generation abilities suggested in the Wowhead guide.

Update the documentation at the top of the priority to list each setting
in `/hekili` > *Protection* and my recommendation for the setting.