fix: update blood death knight priorities for 11.0.5
Update the BDK priority based on the official guides for TWW 11.0.5 to
follow the Icy Veins guide written by the same authors as the Wowhead
guide. The IV guide has a few more details in the priority than the
Wowhead one.

Update the Kyrasis BDK priority from his 11.0.5 BDK guide.