fix: update Equinox monk priority for 11.0.5
Update to the latest changes to the Equinox brewmaster monk guide for
11.0.5. Spinning Crane Kick now correctly counts toward Flurry Strikes,
but it's still not worth casting due to lack of damage compared to
Tiger's Palm. Rising Sun Kick priority in AoE has slightly changed so
that it's used strictly as a filler. Also support Master of Harmony hero
talents.