fix: suggest Demon Spikes more aggressively
Suggest Demon Spikes even if demon form is ready, but caption those
suggestions with Fel Devastation or Metamorphosis icons to allow the
player to avoid overlapping Demon Spikes with demon form by choice, or
to go ahead and use Demon Spikes and delay demon form.

Be more consistent with suggesting Demon Spikes in the opener when it's
worthwhile to have the extra armor and parry at the start of a pull,
even if demon form will be used.