fix: improve suggestions when close to capping Rage
Change the way that we determine if we should dump Rage as a protection
warrior. Instead of checking if the next GCD will hit the Rage cap,
compute the number of seconds until the Rage cap is reached based on the
cooldowns of Rage-generating abilities, and proactively try to dump Rage
with Execute or Revenge if possible, but always falling back to Ignore
Pain.

This should avoid capping on Rage as often due to procs.